Propagate ICMP Host & Network Unreachable errors back to applications so they can immediately take appropriate action instead of waiting for timeouts.
Change outgoing UDP sockets to use connect() & send() so the correct
destination is remembered for the lifetime of a UDP session. Previously when
port forwarding UDP connections only the first packet would be redirected,
which is enough for a DNS request & response, but breaks QUIC.

Close outgoing TCP & UDP sockets if they fail to connect and setup fully.